requirements.txt
Flask==1.1.2
gunicorn==20.0.4
twilio==6.63.2
requests==2.22.0
firebase_admin==4.3.0
datetime==4.3
googletrans==3.1.0a0
validators==0.18.2